refactor: move logger request middleware to middleware package
