+++ middleware/logger_req.go
@@ -1,4 +1,4 @@
-package bastion
+package middleware
 
 import (
 	"net/http"
@@ -8,8 +8,15 @@ import (
 	"github.com/rs/zerolog/hlog"
 )
 
+func getLoggerWithLevel(r *http.Request, status int) *zerolog.Event {
+	if status >= 500 {
+		return hlog.FromRequest(r).Error()
+	}
+	return hlog.FromRequest(r).Info()
+}
+
 // LoggerRequest some provided extra handler to set some request's context fields.
-func LoggerRequest(opts *Options) []func(next http.Handler) http.Handler {
+func LoggerRequest(extraLog bool) []func(next http.Handler) http.Handler {
 	hdls := []func(next http.Handler) http.Handler{}
 
 	access := hlog.AccessHandler(func(r *http.Request, status, size int, duration time.Duration) {
@@ -24,17 +31,10 @@ func LoggerRequest(opts *Options) []func(next http.Handler) http.Handler {
 
 	hdls = append(hdls, access)
 	hdls = append(hdls, hlog.RequestIDHandler("req_id", "Request-Id"))
-	if !opts.isDEV() {
+	if extraLog {
 		hdls = append(hdls, hlog.RemoteAddrHandler("ip"))
 		hdls = append(hdls, hlog.UserAgentHandler("user_agent"))
 		hdls = append(hdls, hlog.RefererHandler("referer"))
 	}
 	return hdls
-}
-
-func getLoggerWithLevel(r *http.Request, status int) *zerolog.Event {
-	if status >= 500 {
-		return hlog.FromRequest(r).Error()
-	}
-	return hlog.FromRequest(r).Info()
 }
